started work on modules going to bed

+++ fitting_engine/src/stats/drone.rs
@@ -4,14 +4,14 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize, Stat, PartialEq)]
 pub struct Drone {
-    pub control_range: f32,
+    pub control_range: u32,
     pub capacity: u16,
     pub bandwidth: u16,
     pub max_drones: u8,
 }
 
 impl Drone {
-    pub fn new(control_range: f32, capacity: u16, bandwidth: u16, max_drones: u8) -> Self {
+    pub fn new(control_range: u32, capacity: u16, bandwidth: u16, max_drones: u8) -> Self {
         Self {
             control_range,
             capacity,
